+++ packages/build-plugin-component-screenshot/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.2.0
 
 - feat: write `build/screenshots.json`
+- feat: enable on cloud environment